+++ tests/testthat/test-check_information.R
@@ -1,5 +1,11 @@
 test_that("multiplication works", {
-    file <- system.file("workspaces", "workspace_test.xml", package = "rjd3workspace")
+    skip_if_not(rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version)
+
+    file <- system.file(
+        "workspaces",
+        "workspace_test.xml",
+        package = "rjd3workspace"
+    )
     jws <- jws_open(file)
 
     # Check if the SA-Item 3 in the SA-Processing 1 exists