+++ NEWS.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+### Changed
+
+* Use `get_java_version()` instead of `.jversion`
 
 ### Added
 
@@ -28,6 +33,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 * `write_calendars()` accepts now multiple calendars #95
+* New JARS related to version [3.6.0](https://github.com/jdemetra/jdplus-main/releases/tag/v3.6.0)
 
 
 ## [3.5.1] - 2025-06-19
@@ -52,6 +58,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 * Simplified the modification of path in `update_path()`
+* New JARS related to version [3.5.1](https://github.com/jdemetra/jdplus-main/releases/tag/v3.5.1)
+
 
 ### Removed
 
@@ -68,6 +76,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Rename the argument `jsa` in `jsai` in `read_sai()`, `.jsai_results()`, `.jsai_jresults()`, `.jsai_name()`, `.jsai_metadata()`, `.jsai_ts_metadata()`, `replace_sa_item()`, `get_raw_data()`, `get_ts()`, `get_comment()` and  `get_priority()`
 * Rename the argument `ref_jsa` in `ref_jsai` in `set_ts_metadata()`
+* New JARS related to version [3.5.0](https://github.com/jdemetra/jdplus-main/releases/tag/v3.5.0)
+
 
 ### Deprecated 
 
@@ -92,7 +102,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Function `.jws_make_copy()` is deprecated. Please use function `jws_make_copy()` instead.
 * Function `.jsap_make_copy()` is deprecated. Please use function `jsap_make_copy()` instead.
 * Function `.jws_compute()` is deprecated. Please use function `jws_compute()` instead.
-* Function `.jsap_sai()` is deprecated. Please use function `jsap_sai()` instead.
 * Function `.jsap_refresh` is deprecated. Please use function `jsap_refresh` instead.
 * Function `.jws_refresh` is deprecated. Please use function `jws_refresh` instead.
 
@@ -124,24 +133,44 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * new function `.jsa_jresults()`, `.jread_sap()` and `.jread_workspace()` to  extract the Java object of the results of a SAItem.
 
+### Changed
+
+* New JARS related to version [3.2.4](https://github.com/jdemetra/jdplus-main/releases/tag/v3.2.4)
+
 
 ## [3.2.3] - 2024-07-12
 
 ### Changed
 
-* new jars
+* New JARS related to version [3.2.3](https://github.com/jdemetra/jdplus-main/releases/tag/v3.2.3)
 
 
 ## [3.2.2] - 2024-03-15
 
+### Changed
+
+* New JARS related to version [3.2.2](https://github.com/jdemetra/jdplus-main/releases/tag/v3.2.2)
+
 
 ## [3.2.1] - 2023-12-12
+
+### Changed
+
+* New JARS related to version [3.2.1](https://github.com/jdemetra/jdplus-main/releases/tag/v3.2.1)
 
 
 ## [3.2.0] - 2023-11-24
 
+### Changed
+
+* New JARS related to version [3.2.0](https://github.com/jdemetra/jdplus-main/releases/tag/v3.2.0)
+
 
 ## [3.1.0] - 2023-08-11
+
+### Added
+
+* New JARS related to version [3.1.0](https://github.com/jdemetra/jdplus-main/releases/tag/v3.1.0)
 
 
 [Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/v3.6.0...HEAD